Move concrete classes out of `tasks/__init__.py`
